+++ mothernode/java/src/com/orbswarm/swarmcomposer/composer/SwarmListener.java
@@ -4,5 +4,5 @@ package com.orbswarm.swarmcomposer.composer;
  * @author Simran Gleason
  */
 public interface SwarmListener {
-    public void updateSwarmDistances(double radius, int nbeasties, int[][] distances);
+    public void updateSwarmDistances(double radius, int nbeasties, double[][] distances);
 }
